+++ lib/sgcop.rb
@@ -5,5 +5,5 @@ require 'sgcop/inject'
 Sgcop::Inject.defaults!
 
 require 'rubocop/cop/sgcop/simple_format'
-require 'rubocop/cop/sgcop/sleep'
 require 'rubocop/cop/sgcop/whenever'
+require 'rubocop/cop/sgcop/capybara/sleep'
