+++ lib/sgcop/version.rb
@@ -1,3 +1,3 @@
 module Sgcop
-  VERSION = '0.1.2'.freeze
+  VERSION = '0.2.0'.freeze
 end
